more class structure refactoring
